requirements.txt
fastapi
PyPDF2
uvicorn
python-dotenv